New calibrators.py
integration_correction is the new set_integration_correction that used to be in mc.py, but it improved to be more accurate. It has been moved to calibrators.py as it should be used for all data sources.
CameraDL1Calibrator is the Component used to fill the dl1 container. It also handles the integration_correction and neighbour list.