Restore masked array support for hillas
In a previous commit the hillas parameters were changed to use the “@“
operator to perform dot matrix multiplication.
This does not seem to handle masked arrays correctly, completely
ignoring the mask.
Instead I suggest to use the np.ma.dot method, which is applicable to
both default numpy arrays and masked arrays.