Creation of FactoryMeta
FactoryMeta enables:
- The creation of the product traitlet automatically, containing a list
of the possible classes the user can choose from cmdline
- The copying of the subclass traitlets automatically to the Factory,
so the user can easily configure the product class, and see useful help
messages for the factory
- The altering of the Factory traitlet help messages to correctly
indicate which traitlets are valid for which products

As a result the Factory classes are extremely simplified. The only
thing required to do to create a Factory is to set a `base` class.