Feature/add gain selection (#673)
add preliminary code to do gain selection

 *  moved pick_gain_channel() from tino_cta repo
 *  Created classes for GainSelector (base class)
    * NullGainSelector (will leave both gains remaining)
    * ThresholdGainSelector (chooses gain based on threshold, defined in a calibration table called gain_selection_table.ecsv in ctapipe-extra)
    * SimpleGainSelector (chooses one channel always)
    * Created a GainSelectorFactory to allow the user to choose the method

The ThresholdGainSelector has an option to either replace the full waveform with the high-gain one, if the threshold is exceeded, or to only replace samples above threshold (the latter assumes all calibration has been applied such that this will not introduce a discontinuity). The default is to replace the full waveform, as done in current experiments.