change test_event scope to session
This speeds up the tests in that the test_event is only loaded once
per pytest run.  The only problem then is that when used, the event
should be deepcopied first so that it doesn't get modified between
tests... Not sure if there is a better way to do that.