Fix CameraGeometry.transform for cam_rotations != 0, fix CameraDisplay pix_rotation for square pixels
Pixel rotation was broken for square pixels, because matplotlib rotates a
plt.Rectangle around a corner, not the center. Fixed by using RegularPolygon
with 4 vertices instead of Rectangle for square pixels.

CameraGeometry.transform_to did not work correctly for cam_rotation != 0.
I think this was not an issue before, as we always applied derotation when
reading simtel files and probably nothing else uses a cam rotation, anyways, fixed now.

I thought this should have been solvable by using the CameraFrame.rotation attributes,
but it seems they are not correctly handled in the transformation or
I did something wrong in the tests.
So for now, we rotate in CameraGeometry.transform_to