fix bug:  metadata fails on systems without `pwd`
The standard library module `pwd` doesn't exist on all systems (e.g.
Windows or WASM), so don't rely on it. This just replaces the call to it
with a default value if it is not there.

This feature was only used to get the user's name if they forgot to
explicitly include it, so it should not break things if it fails (which
it currently does).