Add an ExitStack to Tool
This enables proper handling of context-manager members
inside Tool.run. Tool-members that require a cleanup step
should be implemented as context managers and be added to
the tool like this:
```
self.foo = self.enter_context(Foo())
```
This will ensure that ``Foo.__exit__`` is called when the
``Tool`` terminates, for whatever reason.