Add option to override obs_id in SimTelEventSource
Simulations were performed that re-use the same CORSIKA run numbers,
e.g. different pointing positions for the same particles in prod5
and also different particles for the same pointing.

To still enable unique obs_ids in merged files for these productions,
an "override_obs_id" option is added to the ``SimTelEventSource``
that can be used to assign new, globally unique ids.

The original run number is added to the ``SimulationConfigContainer``.