migration of flatfield and pedestal components from lstchain
This commit adds a migration proposal for the flatfield and pedestal components from lstchain into ctapipe.

Shared functionality is moved into a parent component, while the different subcomponents hold the _process() function for the conventional processing of flatfield and pedestal events.

Additional subcomponents with complex proccesing routines of flatfield and pedestal events can be added in the future.