Fix StereoTrigger non-deterministically discarding LST-1 in prod6 files
This was one of the weirder bugs I have ever tracked down.
The symptom was that in consecutive runs of `ctapipe-process` on the
same input file, sometimes LST-1 was missing from the output completely.

Tracking it down to the SoftwareTrigger class, which used a mixture of
using ``TelescopeDescription`` and its string representation for storing
and looking up things.

The dict `_ids_by_type` was the culprit, its content depended on the
hash of str(TelescopeDescription), which was random, since hash of
strings are random for security reasons in python.