Improve reference metadata handling for EventSource
The recent addition of attaching the reference metadata
of input files to the provenance information implemented
only that the reference metadata is read directly by
ctapipe from the input file. This made it necessary
to either support all possible input file types
or impossible for plugin event sources to provide this
metadata on their own. It also makes the assumption
1 EventSource = 1 input file. This is not true for some
event sources.

This issue is solved by:
* Adding the possibility to directly provide the
  reference metadata to `add_input_file`
* Move the responsibility of calling `add_input_file` from
  the EventSource baseclass to the implementation